move platform utils to localstack.utils.platform
